+++ apps/extension/src/ui/domains/Portfolio/AssetDetails/PopupTokenBalances/BittensorTokenBalanceList.tsx
@@ -38,7 +38,11 @@ export const BittensorTokenBalanceList = ({
   const [fistGroupStake] = groupedStakesByNetuid ?? []
   const { chainOrNetwork, token, detailRows, status, networkType } = tokenBalances
   const { subnetData, isError, isLoading, isFetchingNextPage } = combinedSubnetData
-  const { price_change_1_day } = subnetData[Number(listKey)] ?? {}
+  const {
+    price_change_1_day,
+    subnet_name,
+    symbol: subnetTokenSymbol,
+  } = subnetData[Number(listKey)] ?? {}
 
   // wait for data to load
   if (!chainOrNetwork || !token || balances.count === 0) return null
@@ -47,11 +51,14 @@ export const BittensorTokenBalanceList = ({
   const {
     meta: {
       alphaToTaoRate,
-      dynamicInfo: { subnetIdentity: { subnetName } = {}, tokenSymbol } = {},
+      dynamicInfo: {
+        subnetIdentity: { subnetName = subnet_name } = {},
+        tokenSymbol = subnetTokenSymbol,
+      } = {},
     } = {},
   } = fistGroupStake
 
-  const subnetListName = `${listKey} | ${subnetName || ""} ${tokenSymbol || ""}`.trim()
+  const subnetListName = `${listKey} | ${subnetName} ${tokenSymbol || ""}`.trim()
   const chainName = isRootStake || isChainIfo ? chainOrNetwork.name || "" : subnetListName
 
   const symbol = isRootStake ? token.symbol : tokenSymbol