+++ packages/extension-core/src/domains/accounts/helpers.ts
@@ -5,7 +5,9 @@ import {
   Account,
   getAccountGenesisHash,
   isAccountEthereum,
+  isAccountLedgerPolkadotGeneric,
   isAccountNotContact,
+  isAccountPolkadot,
 } from "@talismn/keyring"
 
 import { getEthDerivationPath } from "../ethereum/helpers"
@@ -44,13 +46,20 @@ export const sortAccounts =
     return sorted
   }
 
+const getInjectedAccountType = (account: Account): InjectedAccount["type"] => {
+  if (isAccountEthereum(account)) return "ethereum"
+  // some dapps pass only sr25519 as filter
+  if (isAccountPolkadot(account)) return "sr25519"
+  throw new Error("Unsupported account type")
+}
+
 export const getPjsInjectedAccount = (
   account: Account,
   options = { includePortalOnlyInfo: false },
 ): InjectedAccount | (InjectedAccount & { readonly: boolean; partOfPortfolio: boolean }) => ({
   address: account.address,
   name: account.name,
-  type: getAccountKeypairType(account),
+  type: getInjectedAccountType(account),
   ...("genesisHash" in account && account.genesisHash ? { genesisHash: account.genesisHash } : {}),
   ...(options.includePortalOnlyInfo
     ? {
@@ -115,5 +124,6 @@ export const isCurveCompatibleWithChain = (
 export const isAccountCompatibleWithChain = (chain: Chain, account: Account) => {
   const genesisHash = getAccountGenesisHash(account)
   if (genesisHash && genesisHash !== chain.genesisHash) return false
+  if (isAccountLedgerPolkadotGeneric(account) && !chain.hasCheckMetadataHash) return false
   return isAccountEthereum(account) ? chain.account === "secp256k1" : chain.account !== "secp256k1"
 }